requirement.txt
langchain
langchain-openai
langchain-community
langchain-core
flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.2.1
PyPDF2==3.0.1
pydub==0.25.1
websocket-client==1.9.0
websockets
redis==5.2.0
pinecone-client==3.0.0
assemblyai==0.16.0
google-cloud-texttospeech==2.16.1
phonemizer==3.3.0
tiktoken>=0.7.0,<1.0.0


